Migrate pairing flow tests to playwright
Add utility function for waiting until a locator is visible
The isVisible function does not wait for the element to become visible and returns immediately.
