+++ frontend/src/e2e-playwright/utils/index.ts
@@ -2,10 +2,11 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import { isEqual } from 'lodash'
 import { differenceInSeconds } from 'date-fns'
 import config from 'e2e-test-common/config'
+import { isEqual } from 'lodash'
 import { BaseError } from 'make-error-cause'
+import { Locator } from 'playwright'
 
 /**
  * Returns a promise that is resolved after the given amount of milliseconds
@@ -107,6 +108,17 @@ export async function waitUntilFalse(f: () => Promise<boolean>) {
     f,
     (value) => !value,
     (value) => expect(value).toStrictEqual(false)
+  )
+}
+
+/**
+ * Waits until the given function returns a promise which resolves to false
+ */
+export async function waitUntilVisible(locator: Locator) {
+  return waitForCondition(
+    () => locator.isVisible(),
+    (value) => value,
+    (value) => expect(value).toStrictEqual(true)
   )
 }
